+++ my-nextjs-app/src/app/components/input-form.tsx
@@ -1,8 +1,7 @@
 "use client";
 import { useState } from "react";
 import { Button } from "@/components/ui/button";
-import { CheckCircle2, Loader2, PlusCircle, Trash2 } from "lucide-react";
-import { Input } from "@/components/ui/input";
+import { CheckCircle2, Loader2, PlusCircle } from "lucide-react";
 import { Item } from "../types";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 import {
@@ -14,8 +13,9 @@ import {
   CardTitle,
 } from "@/components/ui/card";
 import { promptFormat } from "../constants";
+import { SelectForm } from "./select-form";
 
-export default function InputForm() {
+export const InputForm = () => {
   const [items, setItems] = useState<Item[]>([{ id: 1, value: "" }]);
   const [selected, setSelected] = useState<boolean>(false);
   const [result, setResult] = useState<string>();
@@ -26,14 +26,6 @@ export default function InputForm() {
     const newId =
       items.length > 0 ? Math.max(...items.map((item) => item.id)) + 1 : 1;
     setItems([...items, { id: newId, value: "" }]);
-  };
-
-  const deleteItem = (id: number) => {
-    setItems(items.filter((item) => item.id !== id));
-  };
-
-  const updateItem = (id: number, value: string) => {
-    setItems(items.map((item) => (item.id === id ? { ...item, value } : item)));
   };
 
   const handleSubmit = async (e: React.FormEvent) => {
@@ -83,30 +75,7 @@ export default function InputForm() {
         <CardContent>
           {!selected ? (
             <form onSubmit={handleSubmit} className="space-y-4">
-              <div className="space-y-2">
-                {items.map((item, index) => (
-                  <div key={item.id} className="flex items-center space-x-2">
-                    <Input
-                      type="text"
-                      value={item.value}
-                      onChange={(e) => updateItem(item.id, e.target.value)}
-                      placeholder={`選択肢 ${index + 1}`}
-                      aria-label={`選択肢 ${index + 1}`}
-                      className="flex-grow"
-                    />
-                    <Button
-                      type="button"
-                      variant="destructive"
-                      size="icon"
-                      onClick={() => deleteItem(item.id)}
-                      aria-label={`選択肢 ${index + 1} を削除`}
-                      className="flex-shrink-0"
-                    >
-                      <Trash2 className="h-4 w-4" />
-                    </Button>
-                  </div>
-                ))}
-              </div>
+              <SelectForm items={items} setItems={setItems} />
               <div className="flex justify-between items-center pt-4">
                 <Button
                   type="button"
@@ -156,4 +125,4 @@ export default function InputForm() {
       </Card>
     </div>
   );
-}
+};
