+++ my-nextjs-app/src/app/page.tsx
@@ -1,4 +1,4 @@
-import InputForm from "./components/input-form";
+import { InputForm } from "./components/input-form";
 
 export default function Home() {
   return (
